Rename linebreak_ascii to linebreak_simple
It's not just strict ascii
