+++ internal/core/textlayout/linebreak_simple.rs
